+++ ipl/src/main/java/com/backend/Application.java
@@ -1,4 +1,4 @@
-package com.backend.config;
+package com.backend;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
